all API show tests moved over
